Add a button for adding a new option
