+++ utilities/singleFormFieldUtils.tsx
@@ -1,6 +1,6 @@
-import { filter, get } from "lodash-es";
+import { filter, forEach, get } from "lodash-es";
 import formFieldConfig from "../formFieldConfig.json";
-import { SingleFormField, Collection } from "../types";
+import { SingleFormField, Collection, FormFieldGroup } from "../types";
 
 export function calculateCurrentAttributesToDisplay(question: SingleFormField) {
   const currentTypeConfig = filter(formFieldConfig, (entry) => {
@@ -27,10 +27,6 @@ export function updateCollection(
     ["intakeQuestions", questionIdx],
     {}
   );
-  let keyValuePairsForNewQuestionType = {}; // @TODO populate with missing keys and nullish values for the current Type (use case: say a user changes the question type from checkbox to autocomplete and is missing the options section)
-  // if (questionKey === "type") {
-
-  // }
   const modifiedQuestion = {
     ...targetQuestion,
     [questionKey]: newVal,
@@ -39,7 +35,19 @@ export function updateCollection(
     collection?.intakeQuestions || [];
   newIntakeQuestionSet[questionIdx] = modifiedQuestion;
 
-  console.log("deleteMe newIntakeQuestionSet is: ");
-  console.log(newIntakeQuestionSet);
   setCollection({ ...collection, intakeQuestions: newIntakeQuestionSet });
+}
+
+export function updateOptionFormFieldGroupWithOptionList(
+  options: string[],
+  optionFormFieldGroup: FormFieldGroup
+) {
+  forEach(options, (option, optionIdx) => {
+    const newActualValue: {} = { ["Option " + (optionIdx + 1)]: option };
+    optionFormFieldGroup?.setValues
+      ? optionFormFieldGroup.setValues((prevState: {}) => {
+          return { ...prevState, ...newActualValue };
+        })
+      : undefined;
+  });
 }
